Comparison: Z/32Z with sigma as u32 instead of u128

+++ src/stream.rs
@@ -12,7 +12,7 @@ pub trait FSSKey: Sized {
 
     unsafe fn to_raw_line(&self, raw_line_pointer: *mut u8);
 
-    fn eval(&self, prg: &mut impl PRG, party_id: u8, x: u32) -> i8;
+    fn eval(&self, prg: &mut impl PRG, party_id: u8, x: u32) -> u32;
 
     fn generate_keypair(prg: &mut impl PRG) -> (Self, Self);
 }
